cron-sync/requirements.txt
pyodbc
python-dotenv
mysql-connector-python